refactor: creating BasePolynomial abstract class
intended for future implementations of matrix polynomials
